@@ -10,7 +10,7 @@
   "dependencies": {
     "ant-design-vue": "^1.7.4",
     "axios": "^0.21.0",
-    "vue": "^2.6.11",
+    "vue": "^3.0.7",
     "vue-router": "^3.2.0",
     "vuex": "^3.4.0"
   },
@@ -24,6 +24,6 @@
     "eslint": "^6.7.2",
     "eslint-plugin-vue": "^6.2.2",
     "typescript": "~4.1.5",
-    "vue-template-compiler": "^2.6.11"
+    "@vue/compiler-sfc": "^3.0.7"
   }
 }
